Show cancel modal when clicking on account settings
