+++ packages/datatrak-web/src/layout/Header/LogoButton.tsx
@@ -5,28 +5,10 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
 import { Link as RouterLink, useMatch } from 'react-router-dom';
-import { Button, PageContainer, CancelConfirmModal } from '../components';
-import { HEADER_HEIGHT, ROUTES } from '../constants';
-import { UserMenu } from './UserMenu';
+import { Button, CancelConfirmModal } from '../../components';
+import { HEADER_HEIGHT, ROUTES } from '../../constants';
 
-const Wrapper = styled.div`
-  background: ${({ theme }) => theme.palette.background.paper};
-  width: 100%;
-  box-shadow: inset 0 0 1px #333;
-`;
-
-const Container = styled(PageContainer).attrs({
-  maxWidth: false,
-})`
-  position: relative;
-  z-index: 1;
-  height: ${HEADER_HEIGHT};
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-`;
-
-const LogoButton = styled(Button)`
+const Logo = styled(Button)`
   height: ${HEADER_HEIGHT};
   padding: 0;
   background: transparent;
@@ -45,7 +27,7 @@ const LogoButton = styled(Button)`
   }
 `;
 
-export const Header = () => {
+export const LogoButton = () => {
   const [surveyCancelModalIsOpen, setIsOpen] = useState(false);
   const isSurveyScreen = !!useMatch(ROUTES.SURVEY_SCREEN);
   const isSuccessScreen = !!useMatch(ROUTES.SURVEY_SUCCESS);
@@ -62,14 +44,11 @@ export const Header = () => {
   };
 
   return (
-    <Wrapper>
-      <Container>
-        <LogoButton component={RouterLink} onClick={onClickLogo} to="/" title="Home">
-          <img src="/datatrak-logo-black.svg" alt="tupaia-logo" />
-        </LogoButton>
-        <UserMenu />
-      </Container>
+    <>
+      <Logo component={RouterLink} onClick={onClickLogo} to="/" title="Home">
+        <img src="/datatrak-logo-black.svg" alt="tupaia-logo" />
+      </Logo>
       <CancelConfirmModal isOpen={surveyCancelModalIsOpen} onClose={onClose} />
-    </Wrapper>
+    </>
   );
 };
