+++ packages/datatrak-web/src/layout/UserMenu/MenuList.tsx
@@ -3,13 +3,15 @@
  *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
  */
 
-import React, { ComponentType, ReactNode } from 'react';
+import React, { ComponentType, ReactNode, useState } from 'react';
 import styled from 'styled-components';
+import { useMatch } from 'react-router';
 import { Link, ListItem } from '@material-ui/core';
 import { Button, RouterLink } from '@tupaia/ui-components';
 import { useCurrentUser } from '../../api';
 import { useLogout } from '../../api/mutations';
 import { ROUTES } from '../../constants';
+import { CancelConfirmModal } from '../../components';
 
 const Menu = styled.ul`
   list-style: none;
@@ -48,6 +50,9 @@ export const MenuList = ({
   children?: ReactNode;
   onCloseMenu: () => void;
 }) => {
+  const [surveyCancelModalIsOpen, setIsOpen] = useState(false);
+  const isSurveyScreen = !!useMatch(ROUTES.SURVEY_SCREEN);
+  const isSuccessScreen = !!useMatch(ROUTES.SURVEY_SUCCESS);
   const { isLoggedIn } = useCurrentUser();
   const { mutate: logout } = useLogout();
 
@@ -63,11 +68,19 @@ export const MenuList = ({
     onCloseMenu();
   };
 
+  const onClickInternalLink = e => {
+    if (isSurveyScreen && !isSuccessScreen) {
+      e.preventDefault();
+      setIsOpen(true);
+    }
+  };
+
   const menuItems = isLoggedIn
     ? [
         {
           label: 'Account settings',
           to: ROUTES.ACCOUNT_SETTINGS,
+          onClick: onClickInternalLink,
         },
         helpCentre,
         {
@@ -79,38 +92,46 @@ export const MenuList = ({
     : [helpCentre];
 
   return (
-    <Menu>
-      {children}
-      {menuItems.map(
-        ({
-          label,
-          to,
-          href,
-          isExternal,
-          onClick,
-          component,
-        }: {
-          label: string;
-          to?: string;
-          href?: string;
-          isExternal?: boolean;
-          onClick?: () => void;
-          component?: ComponentType<any> | string;
-        }) => (
-          <MenuListItem key={label} button>
-            <MenuButton
-              component={component || RouterLink}
-              underline="none"
-              target={isExternal ? '_blank' : null}
-              onClick={onClick || onCloseMenu}
-              to={to}
-              href={href}
-            >
-              {label}
-            </MenuButton>
-          </MenuListItem>
-        ),
-      )}
-    </Menu>
+    <>
+      {' '}
+      <Menu>
+        {children}
+        {menuItems.map(
+          ({
+            label,
+            to,
+            href,
+            isExternal,
+            onClick,
+            component,
+          }: {
+            label: string;
+            to?: string;
+            href?: string;
+            isExternal?: boolean;
+            onClick?: (e?: any) => void;
+            component?: ComponentType<any> | string;
+          }) => (
+            <MenuListItem key={label} button>
+              <MenuButton
+                component={component || RouterLink}
+                underline="none"
+                target={isExternal ? '_blank' : null}
+                onClick={onClick || onCloseMenu}
+                to={to}
+                href={href}
+              >
+                {label}
+              </MenuButton>
+            </MenuListItem>
+          ),
+        )}
+      </Menu>
+      <CancelConfirmModal
+        isOpen={surveyCancelModalIsOpen}
+        onClose={() => setIsOpen(false)}
+        confirmLink={ROUTES.ACCOUNT_SETTINGS}
+      />
+    </>
   );
 };
